* commit '236ab1f11ae3beaa9fee4c01e8f541429cd0bf62':
  Use longs instead of ints to store pointers in NIO buffers.

+++ luni/src/main/java/java/nio/Buffer.java
@@ -87,7 +87,7 @@ public abstract class Buffer {
      * This is set in the constructor.
      * TODO: make this final at the cost of loads of extra constructors? [how many?]
      */
-    int effectiveDirectAddress;
+    long effectiveDirectAddress;
 
     /**
      * For direct buffers, the underlying MemoryBlock; null otherwise.
